Changes: 1. Animated enemy prefab
Notes:
- Few more enemy adjustments to make soon
