+++ Assets/Characters/Enemies/KnightEnemy/knight_attack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a98c4492ad4622e4990ffb06155a3e1f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
