+++ Assets/Characters/Enemies/KnightEnemy/knight_death.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 33ce16d7410095448819e40bd83f5c14
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
