+++ Assets/Characters/Enemies/KnightEnemy/knight_hit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 86362e4f031c8f34fabd584e24f89699
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
