+++ Assets/Characters/Enemies/KnightEnemy/knight_idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce30b7eb777cb1241b14dfda254265b8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
