+++ Assets/Characters/Enemies/KnightEnemy/knight_run.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 067302ee58a6825429e0b46fdbe916d9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
