+++ Assets/Characters/Enemies/KnightEnemy/knight_walk.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d6d1b8feb542cb42925ad44674edd72
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
