+++ Assets/Characters/Player/AC_Player.controller
@@ -332,7 +332,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1888cda1b8084f740b275bb6189981af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  boolName: canMove
+  boolName: CanMove
   updateOnState: 0
   updateOnStateMachine: 1
   valueOnEnter: 0
@@ -462,7 +462,7 @@ AnimatorController:
     m_DefaultInt: 0
     m_DefaultBool: 0
     m_Controller: {fileID: 9100000}
-  - m_Name: canMove
+  - m_Name: CanMove
     m_Type: 4
     m_DefaultFloat: 0
     m_DefaultInt: 0
