--- Assets/Knight.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aab4448c94a025b4e89ac920ffbc030a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
