+++ Assets/Scripts/AnimationStrings.cs
@@ -16,6 +16,7 @@ internal class AnimationStrings
     internal static string isOnWall = "IsOnWall";
     internal static string isOnCeiling = "IsOnCeiling";
     internal static string attackTrigger = "attack";
-    internal static string canMove = "canMove";
+    internal static string canMove = "CanMove";
+    internal static string hasTarget = "HasTarget";
 }
