+++ Assets/Scripts/DetectionZone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 75c3b13cb8ad8b744a57d0f3c35f6a1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
