+++ Builds/PC/Prototype Platformer_Data/boot.config
@@ -3,4 +3,4 @@ gfx-enable-native-gfx-jobs=1
 wait-for-native-debugger=0
 hdr-display-enabled=0
 gc-max-time-slice=3
-build-guid=b70aaa9d63144b6f88875ee33454abe7
+build-guid=e213e144d04d4b169a1170d9703e19f1
